django-based-api/requirements.txt
--extra-index-url https://pypi.fury.io/mohbenaicha/ 
book-recommender==0.0.1
uvicorn==0.16.0
django==3.2.9 
pyngo==1.1.0
djangorestframework==3.13.1
python-multipart==0.0.5
typing_extensions>=3.7.4.3
pydantic==1.8.2
loguru==0.5.3

